delete file in Storage
